Fix some errors.
Reworked how character IDs work because LUA doesn't guarantee table orders so we can't
be sure the n'th entry is always the same character (and it wasn't).

If an account cleared their WDB folder, or multiple accounts with different
WDBs share the same Altoholic DB, we might encounter items that the account doesn't have
cached, which makes GetItemInfo return nil, which threw errors. Attempt to load the
item instead by adding it to a tooltip; see
https://www.wowinterface.com/forums/showthread.php?t=34086.

Don't try to show more than 10 character columns in equipment and reputation because
it leads to crashes and we don't have the space anyway.